requirements.txt
django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
django-cleanup==7.0.0
djoser==2.1.0
drf-yasg==1.21.5
django-filter==23.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-smart-selects==1.6.0
sorl-thumbnail==12.9.0